requirements.txt
hvac
psycopg2-binary
pymongo
requests
python-dotenv
bandit
pytest
flake8
python-telegram-bot
matrix-nio
crewai
fastapi
uvicorn
rich
sqlalchemy
<<<<<<< HEAD
sqlcipher3
=======
tenacity==8.5.0
PyYAML==6.0.2
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
>>>>>>> a6084cc3ed82e7829e4008fdba7650ce580d27d4
